Added Bot Play
Added a bot to play against, so you actually have some competition.
Notes: You do not see what the bot rolls (other than when the bot needs to re-roll), this is because I feel the Console gets cluttered if it did. So when the bot rolls the dice, it will either say it needs a re-roll, or give you the score.